refactor(engine): split engine.py into base + helpers + registry
Closes #57. engine.py is now a thin re-export shim (27 LOC); real code
lives in engine_base (ImageEngine + preflight_check), engine_helpers
(EngineCapabilities, errors, capability detection, warn, TwoPhaseMixin,
quantize_transformer), and engine_registry (factory). All four modules
now fit under the 300-LOC file-length gate; the engine.py exemption
entry is removed.
