+++ src/imagecli/engine_base.py
@@ -1,0 +1,298 @@
+"""Abstract :class:`ImageEngine` + :func:`preflight_check`. Helpers/registry are siblings."""
+
+from __future__ import annotations
+
+import gc
+from abc import ABC, abstractmethod
+from collections.abc import Callable
+from pathlib import Path
+from typing import ClassVar
+
+from imagecli import engine_helpers as _h
+from imagecli.engine_helpers import (
+    EngineCapabilities,
+    InsufficientResourcesError,
+    MIN_FREE_RAM_GB,
+    TwoPhaseMixin,
+    logger,
+    quantize_transformer,
+)
+
+
+class ImageEngine(TwoPhaseMixin, ABC):
+    name: str
+    description: str
+    model_id: str
+    vram_gb: float  # approximate minimum VRAM
+    capabilities: ClassVar[EngineCapabilities] = EngineCapabilities()
+
+    def __init__(
+        self,
+        *,
+        compile: bool = True,
+        lora_path: str | None = None,
+        lora_scale: float = 1.0,
+        trigger: str | None = None,
+        embedding_path: str | None = None,
+    ) -> None:
+        self._pipe: object | None = None
+        self._compile = compile
+        self._compiled = False
+        self.lora_path = lora_path
+        self.lora_scale = lora_scale
+        self.trigger = trigger
+        self.embedding_path = embedding_path
+
+    @abstractmethod
+    def _load(self) -> None:
+        """Load the model pipeline into self._pipe. Called lazily on first generate()."""
+
+    def effective_steps(self, requested: int) -> int:
+        """Return the step count that will actually run. Override for engines with fixed steps."""
+        return requested
+
+    def _build_pipe_kwargs(
+        self,
+        prompt: str,
+        *,
+        negative_prompt: str,
+        width: int,
+        height: int,
+        steps: int,
+        guidance: float,
+        generator: object | None,
+    ) -> dict:
+        """Return kwargs for the pipeline call. Override for engine-specific args."""
+        return {
+            "prompt": prompt,
+            "width": width,
+            "height": height,
+            "num_inference_steps": steps,
+            "guidance_scale": guidance,
+            "generator": generator,
+        }
+
+    def generate(
+        self,
+        prompt: str,
+        *,
+        negative_prompt: str = "",
+        width: int = 1024,
+        height: int = 1024,
+        steps: int = 50,
+        guidance: float = 4.0,
+        seed: int | None = None,
+        output_path: Path,
+        callback: Callable[..., dict] | None = None,
+        **kwargs,
+    ) -> Path:
+        """Generate an image and save it to output_path. Returns the saved path."""
+        import random
+
+        import torch
+
+        self._load()
+
+        # Auto-generate seed when none provided — ensures reproducibility
+        if seed is None:
+            seed = random.randint(0, 2**32 - 1)
+        generator = torch.Generator("cuda").manual_seed(seed)
+
+        pipe_kwargs = self._build_pipe_kwargs(
+            prompt,
+            negative_prompt=negative_prompt,
+            width=width,
+            height=height,
+            steps=steps,
+            guidance=guidance,
+            generator=generator,
+        )
+        if callback is not None:
+            pipe_kwargs["callback_on_step_end"] = callback
+
+        if torch.cuda.is_available():
+            torch.cuda.reset_peak_memory_stats()
+
+        with torch.inference_mode():
+            result = self._pipe(**pipe_kwargs)
+
+        image = result.images[0]
+        return self._save_image(
+            image,
+            output_path,
+            seed=seed,
+            steps=steps,
+            guidance=guidance,
+            width=width,
+            height=height,
+        )
+
+    def _save_image(
+        self,
+        image,
+        output_path: Path,
+        *,
+        seed: int,
+        steps: int,
+        guidance: float,
+        width: int,
+        height: int,
+    ) -> Path:
+        """Save image with PNG metadata for reproducibility."""
+        output_path.parent.mkdir(parents=True, exist_ok=True)
+        from PIL.PngImagePlugin import PngInfo
+
+        png_meta = PngInfo()
+        png_meta.add_text("seed", str(seed))
+        png_meta.add_text("engine", self.name)
+        png_meta.add_text("steps", str(steps))
+        png_meta.add_text("guidance", str(guidance))
+        png_meta.add_text("width", str(width))
+        png_meta.add_text("height", str(height))
+        image.save(str(output_path), pnginfo=png_meta)
+        return output_path
+
+    def _finalize_load(self, pipe: object) -> None:
+        """Move pipeline to GPU and apply optimizations. Raises if VRAM is insufficient."""
+        import torch
+
+        if torch.cuda.is_available():
+            free_vram_gb = torch.cuda.mem_get_info(0)[0] / 1024**3
+            if free_vram_gb < self.vram_gb:
+                raise InsufficientResourcesError(
+                    f"Engine {self.name!r} needs ~{self.vram_gb:.1f} GB VRAM, "
+                    f"but only {free_vram_gb:.1f} GB is free. "
+                    f"Close other GPU processes and retry."
+                )
+            pipe.to("cuda")
+        self._optimize_pipe(pipe)
+
+    def _optimize_pipe(self, pipe: object, *, compile: bool | None = None) -> None:
+        """Apply performance optimizations. compile=False overrides self._compile.
+
+        Use compile=False when CPU offload hooks are active, since torch.compile
+        is incompatible with enable_model_cpu_offload.
+        """
+        import torch
+
+        # TF32 matmul — tensor cores on Blackwell (~10-15% speedup, once per process)
+        if not _h._tf32_set:
+            torch.set_float32_matmul_precision("high")
+            _h._tf32_set = True
+
+        # VAE optimizations — reduce peak VRAM for large images and batches
+        if hasattr(pipe, "vae"):
+            pipe.vae.enable_tiling()
+            pipe.vae.enable_slicing()
+
+        # torch.compile — 20-40% speedup after first-run warmup
+        should_compile = self._compile if compile is None else compile
+        if should_compile and not self._compiled:
+            compilable = None
+            if hasattr(pipe, "transformer"):
+                compilable = pipe.transformer
+            elif hasattr(pipe, "unet"):
+                compilable = pipe.unet
+            if compilable is not None:
+                pipe_attr = "transformer" if hasattr(pipe, "transformer") else "unet"
+                logger.info(
+                    "Compiling %s with torch.compile (first run will be slower)...", pipe_attr
+                )
+                setattr(
+                    pipe,
+                    pipe_attr,
+                    torch.compile(compilable, mode="max-autotune-no-cudagraphs"),
+                )
+                self._compiled = True
+
+    def _quantize_transformer(self, pipe: object, sm: tuple[int, int]) -> str:
+        """Quantize the transformer: fp8 on sm≥(8,9), int8 on Ampere."""
+        return quantize_transformer(pipe, sm)
+
+    def clear_cache(self) -> None:
+        """Clear CUDA cache between generations without unloading the model."""
+        import torch
+
+        if torch.cuda.is_available():
+            torch.cuda.empty_cache()
+        gc.collect()
+
+    def cleanup(self) -> None:
+        """Fully unload model and free all VRAM / RAM."""
+        import torch
+
+        if self._pipe is not None:
+            del self._pipe
+            self._pipe = None
+            self._compiled = False
+        if torch.cuda.is_available():
+            torch.cuda.empty_cache()
+        gc.collect()
+
+    def _apply_pivotal_embeddings(self) -> None:
+        """Load and apply pivotal tuning embeddings onto ``self._pipe``.
+
+        No-op when neither ``lora_path`` nor ``embedding_path`` is set. Must be
+        called AFTER the LoRA has been fused/unloaded and BEFORE the
+        transformer is quantized or moved to GPU.
+        """
+        if not (self.lora_path or self.embedding_path):
+            return
+        from imagecli.pivotal import (
+            _patch_encode_prompt,
+            apply_pivotal_to_pipe,
+            load_pivotal_embedding,
+        )
+
+        pivotal = load_pivotal_embedding(
+            self.lora_path,
+            self.trigger,
+            embedding_path=self.embedding_path,
+            te_hidden_size=self._pipe.text_encoder.config.hidden_size,
+        )
+        if pivotal is not None:
+            apply_pivotal_to_pipe(self._pipe, pivotal)
+            _patch_encode_prompt(self._pipe)
+
+
+def preflight_check(engine: ImageEngine) -> None:
+    """Abort early if the system can't safely run this engine. Skipped if already loaded."""
+    if engine._pipe is not None:
+        return
+
+    import torch
+
+    # --- VRAM check ---
+    if torch.cuda.is_available():
+        free_vram, total_vram = torch.cuda.mem_get_info(0)
+        free_vram_gb = free_vram / 1024**3
+        if free_vram_gb < engine.vram_gb:
+            raise InsufficientResourcesError(
+                f"Engine {engine.name!r} needs ~{engine.vram_gb:.1f} GB VRAM, "
+                f"but only {free_vram_gb:.1f} GB / {total_vram / 1024**3:.1f} GB is free. "
+                f"Close other GPU processes (e.g. ollama) and retry."
+            )
+    else:
+        raise InsufficientResourcesError(
+            "No CUDA GPU detected. imagecli requires a CUDA-capable GPU."
+        )
+
+    # --- System RAM check ---
+    # GGUF engines load model weights into system RAM via enable_model_cpu_offload().
+    # Require at least engine.vram_gb of free RAM as a proxy for model size.
+    ram_needed_gb = max(MIN_FREE_RAM_GB, engine.vram_gb)
+    try:
+        with open("/proc/meminfo") as f:
+            for line in f:
+                if line.startswith("MemAvailable:"):
+                    available_kb = int(line.split()[1])
+                    available_gb = available_kb / 1024 / 1024
+                    if available_gb < ram_needed_gb:
+                        raise InsufficientResourcesError(
+                            f"Only {available_gb:.1f} GB system RAM available, "
+                            f"need at least {ram_needed_gb:.1f} GB for engine "
+                            f"{engine.name!r}. Close other applications and retry."
+                        )
+                    break
+    except FileNotFoundError:
+        pass  # non-Linux, skip RAM check
